+++ src/k_fold_cross_validation.h
@@ -24,9 +24,9 @@ namespace MachineLearning
 	{
 
 		public:
-			static std::pair<float, float> run(Predictor& _predictor, std::vector<Instance>& _pool, size_t foldsCount);
+			static std::pair<double, double> run(Predictor& _predictor, std::vector<Instance>& _pool, size_t foldsCount);
 
-			static std::pair<float, float> test(Predictor* _predictor, Pool& _pool, size_t folds, std::string testing_category_name, const std::string& outdir, bool print = false);
+			static std::pair<double, double> test(Predictor* _predictor, Pool& _pool, size_t folds, std::string testing_category_name, const std::string& outdir, bool print = false);
 
 		private:
